+++ lib/src/integrations/apple/apple_command.dart
@@ -1,6 +1,5 @@
 import 'package:args/command_runner.dart';
 import 'package:impaktfull_cli/src/core/command/command/root_command.dart';
-import 'package:impaktfull_cli/src/integrations/apple/certificate/command/apple_certificate_root_command.dart';
 import 'package:impaktfull_cli/src/integrations/apple/provisioning_profile/command/apple_provisioning_profile_root_command.dart';
 
 class AppleRootCommand extends RootCommand {
@@ -17,7 +16,6 @@ class AppleRootCommand extends RootCommand {
   @override
   List<Command<dynamic>> getSubCommands() {
     return [
-      AppleCertificateRootCommand(processRunner: processRunner),
       AppleProvisioningProfileRootCommand(processRunner: processRunner),
     ];
   }