+++ lib/src/integrations/apple/certificate/model/remove/apple_certificate_remove_config_data.dart
@@ -1,9 +1,0 @@
-import 'package:impaktfull_cli/src/core/command/config/command_config.dart';
-
-class AppleCertificateRemoveConfigData extends ConfigData {
-  final String keyChainName;
-
-  const AppleCertificateRemoveConfigData({
-    required this.keyChainName,
-  });
-}